% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexconn_conv_chain.R
\name{flexconn_conv_chain}
\alias{flexconn_conv_chain}
\title{FLEXCONN Convolution Chain}
\usage{
flexconn_conv_chain(object, ds = 2, num_filters = 128,
  kernel_size_1 = 3, kernel_size_2 = 5, prefix = NULL, ndim)
}
\arguments{
\item{object}{Model or layer object}

\item{ds}{Some fixed number???}

\item{num_filters}{Nubmer of filters}

\item{kernel_size_1}{Size of first kernel}

\item{kernel_size_2}{Size of second kernel}

\item{prefix}{output prefix for each layer}

\item{ndim}{Number of dimensions for convolution chain}
}
\value{
Output model
}
\description{
FLEXCONN Convolution Chain
}
\examples{
library(tensorflow)
 t1_input <- layer_input(shape = shape(NULL, NULL, 1))
 t1 <- t1_input \%>\%
 flexconn_conv_chain(
   prefix = "t1"
 )

 t1_3d_input <- layer_input(shape = shape(NULL, NULL, NULL, 1))
 t1_3d <- t1_input \%>\%
 flexconn_conv_chain(
   prefix = "t1"
 )
}
