% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexconn_combined_model.R
\name{flexconn_model}
\alias{flexconn_model}
\title{FLEXCONN Model}
\usage{
flexconn_model(
  prefixes = c("t1", "fl"),
  ndim = 2,
  ds = 2,
  num_filters = 128,
  kernel_size_1 = 3,
  kernel_size_2 = 5
)
}
\arguments{
\item{prefixes}{The prefixes for each of the layers of the model.}

\item{ndim}{Number of dimensions for convolution chain (2 or 3)}

\item{ds}{Some fixed number???}

\item{num_filters}{Nubmer of filters}

\item{kernel_size_1}{Size of first kernel}

\item{kernel_size_2}{Size of second kernel}
}
\value{
A keras model object
}
\description{
FLEXCONN Model
}
\examples{
user = Sys.getenv("USER")
if (user \%in\% c("johnmuschelli", "travis") &
Sys.info()["sysname"] == "Darwin"){
reticulate::use_python(paste0(
"/Library/Frameworks/Python.framework/Versions/3.5/bin/python3"))
} else {
python = system("which python", intern = TRUE)
print(python)
reticulate::use_python(python)
}
res = flexconn_model()
res = flexconn_model(prefixes = c("t1", "fl", "t2"))
}
