% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexconn_predict.R
\name{flexconn_predict}
\alias{flexconn_predict}
\title{Predict from FLEXCONN model}
\usage{
flexconn_predict(model, t1, flair, mask = NULL, patchsize,
  verbose = TRUE, ..., batch_size = 1)
}
\arguments{
\item{model}{A keras model object trained}

\item{t1}{3D array or \code{nifti} image}

\item{flair}{3D array or \code{nifti} image}

\item{mask}{binary 3D array or \code{nifti} image, but of
the whole area of prediction, not the ground truth}

\item{patchsize}{Vector of length 2 (or more)}

\item{verbose}{print diagnostic messages}

\item{...}{additional arguments to \code{\link{get_patches}}}

\item{batch_size}{Size of batches for prediction.
Integer. Passed to
\code{\link[keras]{predict.keras.engine.training.Model}}.}
}
\value{
A vector of predictions, based on the indices of the
mask
}
\description{
Predict from FLEXCONN model
}
\note{
If \code{mask = NULL}, a mask will be generated for
\code{t1 > 0}.
}
