% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matsize.R
\name{get_matsize}
\alias{get_matsize}
\title{Get Matrix Size to Input to Model}
\usage{
get_matsize(num_patches, patchsize)
}
\arguments{
\item{num_patches}{Number of Patches}

\item{patchsize}{Patch Size}
}
\value{
A vector
}
\description{
Get Matrix Size to Input to Model
}
\examples{
get_matsize(5, c(35, 35))
}
