% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_patches.R
\name{get_patches}
\alias{get_patches}
\title{Get Patches from T1 and FLAIR image}
\usage{
get_patches(t1, flair, mask = NULL, patchsize, pad = TRUE,
  normalize = TRUE, verbose = TRUE, only_patches = TRUE)
}
\arguments{
\item{t1}{3D array or \code{nifti} image}

\item{flair}{3D array or \code{nifti} image}

\item{mask}{binary 3D array or \code{nifti} image}

\item{patchsize}{Vector of length 2 (or more)}

\item{pad}{Run \code{\link{pad_image}} on the image
before getting patches (pads then normalizes if \code{normalize = TRUE})}

\item{normalize}{Run \code{\link{normalize_image}} on the image
before getting patches}

\item{verbose}{print diagnostic messages}

\item{only_patches}{Only return the patches, not additional information.}
}
\value{
A list of T1, FLAIR, and Mask Patches
}
\description{
Get Patches from T1 and FLAIR image
}
\note{
If \code{mask = NULL}, a mask will be created based on
voxels greater than the 75th percentile of the FLAIR image.
}
\examples{
library(neurobase)
fname = system.file("extdata", "MPRAGE.nii.gz", package = "flexconn")
t1 = readnii(fname)
fname = system.file("extdata", "FLAIR.nii.gz", package = "flexconn")
flair = readnii(fname)
mask = t1 > 0
patchsize = c(3, 3)
verbose = TRUE
patch = get_patches(t1, flair, mask, patchsize = patchsize)

rm(patch)
}
