% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_flexconn.R
\name{install_flexconn}
\alias{install_flexconn}
\title{Install FLEXCONN Requirements}
\usage{
install_flexconn(..., extra_packages = NULL)
}
\arguments{
\item{...}{arguments to pass to \code{\link[reticulate]{py_install}}}

\item{extra_packages}{Additional packages to install}
}
\value{
Nothing
}
\description{
Install FLEXCONN Requirements
}
\examples{
\dontrun{
install_flexconn()
}
}
