% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_indices.R
\name{mask_indices}
\alias{mask_indices}
\title{Get Mask Indices}
\usage{
mask_indices(mask, seed = NULL, verbose = TRUE)
}
\arguments{
\item{mask}{binary 3D array or \code{nifti} image}

\item{seed}{Seed for random sampling of indices.  If \code{NULL},
no sampling is done}

\item{verbose}{print diagnostic messages}
}
\value{
A 3xV matrix with V number of patches
}
\description{
Get Mask Indices
}
\examples{
dims = rep(10,3)
mask = array(rbinom(prod(dims), size = 1, prob = 0.2),
dim = dims)
}
