% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numpy_percentile.R
\name{numpy_percentile}
\alias{numpy_percentile}
\title{Python Numpy Percentile}
\usage{
numpy_percentile(x, probs = 0.99)
}
\arguments{
\item{x}{a numeric vector}

\item{probs}{Quantile to estimate}
}
\value{
A numeric
}
\description{
Python Numpy Percentile
}
\examples{
numpy_percentile(rnorm(100))
}
