% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_image.R
\name{pad_image}
\alias{pad_image}
\alias{invert_pad_image}
\title{Pad Image}
\usage{
pad_image(vol, padsize)

invert_pad_image(vol, padsize)
}
\arguments{
\item{vol}{3D Array or matrix}

\item{padsize}{Scalar number}
}
\value{
Array with same number of dimensions as \code{vol} but increased
dimensions.
}
\description{
Pad Image
}
\examples{
vol = array(rnorm(10^3), dim = rep(10, 3))
dim(vol)
padded = pad_image(vol, padsize = 2)
dim(padded)
orig = invert_pad_image(padded, padsize = 2)
testthat::expect_equal(orig, vol)
}
