% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches_list.R
\name{patches_list}
\alias{patches_list}
\title{Make a patches list for a population of images}
\usage{
patches_list(t1, flair = NULL, mask, t2 = NULL, patchsize = c(35,
  35), verbose = TRUE, outfile = NULL, write_file = TRUE, ...)
}
\arguments{
\item{t1}{a character vector of filenames for T1 images}

\item{flair}{a character vector of filenames for FLAIR images}

\item{mask}{a character vector of filenames for mask images}

\item{t2}{a character vector of filenames for T2 images}

\item{patchsize}{Vector of length 2 (or more)}

\item{verbose}{print diagnostic messages}

\item{outfile}{Output file to write arrays, an \code{.rds} file.
If this exists, it will be read in versus making patches again}

\item{write_file}{Should an output file be written}

\item{...}{Additional arguments to pass to \code{\link{get_patches}}}
}
\value{
A list of named arrays
}
\description{
Make a patches list for a population of images
}
\examples{
user = Sys.getenv("USER")
if (user \%in\% c("johnmuschelli", "travis") &
Sys.info()["sysname"] == "Darwin"){
  reticulate::use_python(paste0(
   "/Library/Frameworks/Python.framework/Versions/3.5/bin/python3"))
} else {
python = system("which python", intern = TRUE)
print(python)
reticulate::use_python(python)
}
data_dir = "/Volumes/DATA_LOCAL/Projects/ms_lesion_challenge/atlases/none"
  if (dir.exists(data_dir)) {

   i = 1:5
t1 <- paste0("atlas", i, "_T1.nii.gz")
flair = paste0("atlas", i, "_FL.nii.gz")
t2 = NULL
mask = paste0("atlas", i, "_mask.nii.gz")
t1 = file.path(data_dir, t1)
flair = file.path(data_dir, flair)
mask = file.path(data_dir, mask)
verbose = TRUE
patchsize = c(35, 35)
outfile = NULL
write_file = TRUE
}
}
