% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches_to_volume.R
\name{patches_to_volume}
\alias{patches_to_volume}
\alias{get_volume_from_patch}
\title{Convert Patches to a Volume}
\usage{
patches_to_volume(patches, indices, patchsize, verbose = TRUE,
  orig_dim = NULL)

get_volume_from_patch(patches, indices, patchsize, verbose = TRUE,
  invert_pad = TRUE, orig_dim = NULL)
}
\arguments{
\item{patches}{A matrix of patches, generally from
\code{\link{volume_to_patches}}}

\item{indices}{A matrix of 3 by \code{num_patches} indices for the
data to extract}

\item{patchsize}{Vector of length 2 (or more)}

\item{verbose}{print diagnostic messages}

\item{orig_dim}{Original dimensions of volume}

\item{invert_pad}{Run \code{\link{invert_pad_image}} on the image
before returning.}
}
\value{
A 3D volume
}
\description{
Convert Patches to a Volume
}
