% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchsize_to_padsize.R
\name{patchsize_to_padsize}
\alias{patchsize_to_padsize}
\title{Patch Size to Pad Size}
\usage{
patchsize_to_padsize(patchsize)
}
\arguments{
\item{patchsize}{A vector of patch sizes}
}
\value{
A scalar number
}
\description{
Patch Size to Pad Size
}
\examples{
patchsize_to_padsize(c(10, 5))
patchsize_to_padsize(c(5, 5))
}
