% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_small_components.R
\name{remove_small_components}
\alias{remove_small_components}
\title{Remove connected objects with volume <27 voxels}
\usage{
remove_small_components(img, threshold = 0.34)
}
\arguments{
\item{img}{A probability image from \code{\link{flexconn_predict_with_volume}}
or \code{\link{flexconn_predict}}, should be
object of class \code{nifti} or an array}

\item{threshold}{threshold to apply to image to create a mask}
}
\value{
An array or object of class \code{nifti}
}
\description{
Remove connected objects with volume <27 voxels
}
\examples{
user = Sys.getenv("USER")
if (user \%in\% c("johnmuschelli", "travis") &
Sys.info()["sysname"] == "Darwin"){
reticulate::use_python(paste0(
"/Library/Frameworks/Python.framework/Versions/3.5/bin/python3"))
}
set.seed(100)
d = rep(50, 3)
img = array(runif(prod(d), max = 1), dim = d)
cc = remove_small_components(img, threshold = 0.9)
}
