% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_to_patches.R
\name{volume_to_patches}
\alias{volume_to_patches}
\title{Volume to Patches}
\usage{
volume_to_patches(vol, indices, patchsize, verbose = TRUE)
}
\arguments{
\item{vol}{3D array or \code{nifti} image}

\item{indices}{A matrix of 3 by \code{num_patches} indices for the
data to extract}

\item{patchsize}{Vector of length 2 (or more)}

\item{verbose}{print diagnostic messages}
}
\value{
A list of image and mask Patches
}
\description{
Volume to Patches
}
