% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_patches.R
\name{get_patch_from_volume}
\alias{get_patch_from_volume}
\title{Get Patches from 3D Volume}
\usage{
get_patch_from_volume(vol, mask, patchsize, verbose = TRUE)
}
\arguments{
\item{vol}{3D array or \code{nifti} image}

\item{mask}{binary 3D array or \code{nifti} image}

\item{patchsize}{Vector of length 2 (or more)}

\item{verbose}{print diagnostic messages}
}
\value{
A list of image and mask Patches
}
\description{
Get Patches from 3D Volume
}
