% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_patches.R
\name{get_patches}
\alias{get_patches}
\title{Get Patches from T1 and FLAIR image}
\usage{
get_patches(t1, flair, mask, patchsize, verbose = TRUE)
}
\arguments{
\item{t1}{3D array or \code{nifti} image}

\item{flair}{3D array or \code{nifti} image}

\item{mask}{binary 3D array or \code{nifti} image}

\item{patchsize}{Vector of length 2 (or more)}

\item{verbose}{print diagnostic messages}
}
\value{
A list of T1, FLAIR, and Mask Patches
}
\description{
Get Patches from T1 and FLAIR image
}
