% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_image.R
\name{normalize_image}
\alias{normalize_image}
\alias{image_peak}
\alias{flexconn_image_peak}
\title{Normalize Image Based on the Peak}
\usage{
normalize_image(
  vol,
  contrast = c("T1", "PD", "T2", "FLAIR", "FL"),
  peak_estimator = c("Python", "R"),
  verbose = TRUE
)

image_peak(vol, contrast = c("T1", "PD", "T2", "FLAIR", "FL"), verbose = TRUE)

flexconn_image_peak(
  vol,
  contrast = c("T1", "PD", "T2", "FLAIR", "FL"),
  verbose = TRUE
)
}
\arguments{
\item{vol}{3D array or \code{nifti} image}

\item{contrast}{What imaging sequence of MRI is this volume}

\item{peak_estimator}{Which functions to use to estimate peak,
either wrapper Python code or native R code.}

\item{verbose}{print diagnostic messages}
}
\value{
The peak of the image is returned for \code{image_peak}
and the normalized image in \code{normalize_image}
}
\description{
Normalize Image Based on the Peak
}
\examples{
user = Sys.getenv("USER")
if (user \%in\% c("johnmuschelli", "travis") &
Sys.info()["sysname"] == "Darwin"){
reticulate::use_python(paste0(
"/Library/Frameworks/Python.framework/Versions/3.5/bin/python3"))
} else {
python = system("which python", intern = TRUE)
print(python)
reticulate::use_python(python)
}
set.seed(1)
vol = array(rpois(10^3, lambda = 10), dim = rep(10, 3))
vol[ vol < 6] = 0
normed = normalize_image(vol, "T1")

}
