% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numpy_percentile.R
\name{numpy_percentile}
\alias{numpy_percentile}
\title{Python Numpy Percentile}
\usage{
numpy_percentile(x, probs = 0.99)
}
\arguments{
\item{x}{a numeric vector}

\item{probs}{Quantile to estimate}
}
\value{
A numeric
}
\description{
Python Numpy Percentile
}
\examples{
set.seed(11)
n = 45
x =rnorm(n)
x = sort(x)
q = numpy_percentile(x)
user = Sys.getenv("USER")
if (user \%in\% c("johnmuschelli", "travis") &
Sys.info()["sysname"] == "Darwin"){
reticulate::use_python(paste0(
"/Library/Frameworks/Python.framework/Versions/3.5/bin/python3"))
} else {
python = system("which python", intern = TRUE)
print(python)
reticulate::use_python(python)
}
np = reticulate::import("numpy")
q2 = np$percentile(x, 99L)
q2
q2 - q
quant = quantile(x, probs = 0.99)
q2 - quant

np$percentile(sort(as.double(x)), 99L)
np$percentile(sort(x), 99L)

}
